move test cases array
